+++ docs/_data/navigation.yml
@@ -22,7 +22,7 @@ main:
           - title: 中信银行 (CITIC)
             url: /double-entry-generator/providers/banks/citic/
           - title: 农业银行储蓄卡 (ABC Debit)
-            url: /double-entry-generator/providers/banks/abcdebit/
+            url: /double-entry-generator/providers/banks/abc_debit/
           - title: 汇丰银行香港 (HSBC HK)
             url: /double-entry-generator/providers/banks/hsbchk/
           - title: 加拿大银行 (BMO)
