+++ example/abc_debit/example-abc_debit-output.ledger
@@ -1,4 +1,4 @@
-option "title" "abcdebit"
+option "title" "abc_debit"
 option "operating_currency" "CNY"
 
 1970-01-01 open Assets:ABC:DebitCard
