+++ pkg/analyser/abc_debit/abc_debit.go
@@ -1,4 +1,4 @@
-package abcdebit
+package abc_debit
 
 import (
 	"strings"
