+++ test/abc_debit-test-ledger.sh
@@ -12,14 +12,14 @@ mkdir -p "$ROOT_DIR/test/output"
 
 # generate ABC debit bills output in ledger format
 "$ROOT_DIR/bin/double-entry-generator" translate \
-    --provider abcdebit \
-    --config "$ROOT_DIR/example/abcdebit/config.yaml" \
-    --output "$ROOT_DIR/test/output/test-abcdebit-output.ledger" \
-    "$ROOT_DIR/example/abcdebit/example-abcdebit-records.csv"
+    --provider abc_debit \
+    --config "$ROOT_DIR/example/abc_debit/config.yaml" \
+    --output "$ROOT_DIR/test/output/test-abc_debit-output.ledger" \
+    "$ROOT_DIR/example/abc_debit/example-abc_debit-records.csv"
 
 diff -u --color \
-    "$ROOT_DIR/example/abcdebit/example-abcdebit-output.ledger" \
-    "$ROOT_DIR/test/output/test-abcdebit-output.ledger"
+    "$ROOT_DIR/example/abc_debit/example-abc_debit-output.ledger" \
+    "$ROOT_DIR/test/output/test-abc_debit-output.ledger"
 
 if [ $? -ne 0 ]; then
     echo "[FAIL] ABC debit provider output is different from expected output."
